+++ CHANGELOG.md
@@ -79,6 +79,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Admin administration panel stub with sidekiq monitoring and MP version (@mkasztelnik)
 - Render root categories in page footer (@mkasztelnik)
 - Use bootstrap cards to show service offers (@mkasztelnik)
+- Select affiliation on service order configuration page (@mkasztelnik)
 
 ### Changed
 - Upgrade Sprockets gem to avoid CVE-2018-3760 vulnerability (@mkasztelnik)